# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alpha Games API Docs",
        "description": "API for executing Alpha Games gaming e-commerce functionalities",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença API",
            "url": "https://github.com/felipesilva15/alpha-games-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all users",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "parameters": [
                    {
                        "name": "USUARIO_NOME",
                        "in": "query",
                        "description": "User name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "USUARIO_EMAIL",
                        "in": "query",
                        "description": "User e-mail",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "USUARIO_CPF",
                        "in": "query",
                        "description": "User CPF",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "USUARIO_ID": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "USUARIO_NOME": {
                                                "type": "string",
                                                "example": "Username"
                                            },
                                            "USUARIO_EMAIL": {
                                                "type": "string",
                                                "example": "example@example.com"
                                            },
                                            "USUARIO_CPF": {
                                                "type": "string",
                                                "example": "12685963501"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update an user",
                "operationId": "fd4f39f7820d3f3cbc259c5b30fed830",
                "requestBody": {
                    "description": "Data for update user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "USUARIO_NOME": {
                                        "type": "string",
                                        "example": "Username"
                                    },
                                    "USUARIO_EMAIL": {
                                        "type": "string",
                                        "example": "example@example.com"
                                    },
                                    "USUARIO_SENHA": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "USUARIO_CPF": {
                                        "type": "string",
                                        "example": "12685963501"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "USUARIO_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "USUARIO_NOME": {
                                            "type": "string",
                                            "example": "Username"
                                        },
                                        "USUARIO_EMAIL": {
                                            "type": "string",
                                            "example": "example@example.com"
                                        },
                                        "USUARIO_CPF": {
                                            "type": "string",
                                            "example": "12685963501"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registers an user",
                "operationId": "089999749fbdec98922de2f74ba8d540",
                "requestBody": {
                    "description": "Data for creating a new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "USUARIO_NOME": {
                                        "type": "string",
                                        "example": "Username"
                                    },
                                    "USUARIO_EMAIL": {
                                        "type": "string",
                                        "example": "example@example.com"
                                    },
                                    "USUARIO_SENHA": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "USUARIO_CPF": {
                                        "type": "string",
                                        "example": "12685963501"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "USUARIO_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "USUARIO_NOME": {
                                            "type": "string",
                                            "example": "Username"
                                        },
                                        "USUARIO_EMAIL": {
                                            "type": "string",
                                            "example": "example@example.com"
                                        },
                                        "USUARIO_CPF": {
                                            "type": "string",
                                            "example": "12685963501"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get a user by ID",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "USUARIO_ID": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "USUARIO_NOME": {
                                            "type": "string",
                                            "example": "Username"
                                        },
                                        "USUARIO_EMAIL": {
                                            "type": "string",
                                            "example": "example@example.com"
                                        },
                                        "USUARIO_CPF": {
                                            "type": "string",
                                            "example": "12685963501"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/adresses": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all user adresses",
                "operationId": "086b88b73e8103220e07b94b4b4f12dd",
                "responses": {
                    "200": {
                        "description": "User adresses list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Address"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/orders": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all user orders",
                "operationId": "75bd13636229da163281953ad39b5b79",
                "responses": {
                    "200": {
                        "description": "User orders list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDTO"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "EXCPHAND001"
                    },
                    "endpoint": {
                        "type": "string",
                        "example": "api/endpoint"
                    },
                    "message": {
                        "type": "string",
                        "example": "Ocorreu um erro."
                    }
                },
                "type": "object"
            },
...
            "OrderDTO": {
                "properties": {
                    "PEDIDO_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "USUARIO_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "STATUS_ID": {
                        "type": "integer",
                        "example": 2
                    },
                    "PEDIDO_DATA": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-27T03:00:00.000000Z"
                    },
                    "ENDERECO_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "$ref": "#/components/schemas/OrderStatus"
                    }
                },
                "type": "object"
            },
...
            "Address": {
                "properties": {
                    "ENDERECO_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "ENDERECO_NOME": {
                        "type": "string",
                        "example": "Casa"
                    },
                    "ENDERECO_CEP": {
                        "type": "string",
                        "example": "01001000"
                    },
                    "ENDERECO_LOGRADOURO": {
                        "type": "string",
                        "example": "Praça da Sé"
                    },
                    "ENDERECO_NUMERO": {
                        "type": "string",
                        "example": "2589"
                    },
                    "ENDERECO_COMPLEMENTO": {
                        "type": "string",
                        "example": "1º Andar, Apto 4"
                    },
                    "ENDERECO_CIDADE": {
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "ENDERECO_ESTADO": {
                        "type": "string",
                        "example": "SP"
                    },
                    "USUARIO_ID": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "OrderStatus": {
                "properties": {
                    "STATUS_ID": {
                        "type": "integer",
                        "example": 1
                    },
                    "STATUS_DESC": {
                        "type": "string",
                        "example": "Em andamento"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}